Implemented Date formatter
Implemented a date formatter method, though only the old method works. I have yet to debug the issue with the newer method. Also the implementation of the back button and reverse animation remains. As well as adding a button on each search result to fetch further details for the video is planned in future implementation. And will most likely be the final feature for this project.